fix(217-segfault-in-create-lambda-when-started-concurrently): add S3 serer integration tests
