+++ docker/Dockerfile
@@ -1,5 +1,7 @@
 FROM jensvogt/aws-mock-base:latest
 
+RUN apt-get update && apt-get upgrade && apt-get -y install gh
+
 # Install basics
 RUN mkdir -p /usr/src
