+++ dto/CMakeLists.txt
@@ -11,7 +11,8 @@ set(SNS_SOURCES src/sns/SqsNotificationRequest.cpp src/sns/SubscribeRequest.cpp 
 set(S3_SOURCES src/s3/CreateBucketRequest.cpp src/s3/PutBucketVersioningRequest.cpp src/s3/PutObjectRequest.cpp
         src/s3/GetObjectRequest.cpp src/s3/GetObjectResponse.cpp src/s3/DeleteObjectRequest.cpp
         src/s3/DeleteObjectsRequest.cpp src/s3/DeleteObjectsResponse.cpp src/s3/CompleteMultipartUploadResult.cpp
-        src/s3/CopyObjectRequest.cpp src/s3/CopyObjectResponse.cpp)
+        src/s3/CopyObjectRequest.cpp src/s3/CopyObjectResponse.cpp src/s3/CreateBucketConstraint.cpp
+        src/s3/UserIdentity.cpp)
 set(DOCKER_SOURCES src/docker/Port.cpp src/docker/Container.cpp src/docker/ListContainerResponse.cpp
         src/docker/CreateContainerRequest.cpp)
 set(LAMBDA_SOURCES src/lambda/ListTagsResponse.cpp src/lambda/ListFunctionResponse.cpp src/lambda/DeadLetterConfig.cpp
